+++ examples/abp_pcap_detection/run.py
@@ -27,7 +27,7 @@ from morpheus.stages.general_stages import AddClassificationsStage
 from morpheus.stages.general_stages import MonitorStage
 from morpheus.stages.inference.inference_triton import TritonInferenceStage
 from morpheus.stages.input.from_file import FileSourceStage
-from morpheus.stages.output.serialize import SerializeStage
+from morpheus.stages.postprocess.serialize import SerializeStage
 from morpheus.stages.output.to_file import WriteToFileStage
 from morpheus.pipeline.pipeline import LinearPipeline
 from morpheus.stages.preprocess.preprocessing import DeserializeStage
