+++ examples/gnn_fraud_detection_pipeline/run.py
@@ -25,7 +25,7 @@ from morpheus.config import CppConfig
 from morpheus.config import PipelineModes
 from morpheus.stages.general_stages import MonitorStage
 from morpheus.stages.input.from_file import FileSourceStage
-from morpheus.stages.output.serialize import SerializeStage
+from morpheus.stages.postprocess.serialize import SerializeStage
 from morpheus.stages.output.to_file import WriteToFileStage
 from morpheus.pipeline.pipeline import LinearPipeline
 from morpheus.stages.preprocess.preprocessing import DeserializeStage
