+++ tests/test_add_scores_stage_pipe.py
@@ -22,7 +22,7 @@ import pandas as pd
 from morpheus.pipeline import LinearPipeline
 from morpheus.stages.general_stages import AddScoresStage
 from morpheus.stages.input.from_file import FileSourceStage
-from morpheus.stages.output.serialize import SerializeStage
+from morpheus.stages.postprocess.serialize import SerializeStage
 from morpheus.stages.output.to_file import WriteToFileStage
 from morpheus.stages.preprocess.preprocessing import DeserializeStage
 from utils import TEST_DIRS
