+++ tests/test_cli.py
@@ -37,14 +37,14 @@ from morpheus.stages.inference.inference_triton import TritonInferenceStage
 from morpheus.stages.input.from_cloudtrail import CloudTrailSourceStage
 from morpheus.stages.input.from_file import FileSourceStage
 from morpheus.stages.input.from_kafka import KafkaSourceStage
-from morpheus.stages.output.serialize import SerializeStage
+from morpheus.stages.postprocess.serialize import SerializeStage
 from morpheus.stages.output.to_file import WriteToFileStage
 from morpheus.stages.output.to_kafka import WriteToKafkaStage
-from morpheus.stages.output.validation import ValidationStage
+from morpheus.stages.postprocess.validation import ValidationStage
 from morpheus.stages.postprocess.mlflow_drift import MLFlowDriftStage
 from morpheus.stages.postprocess.timeseries import TimeSeriesStage
-from morpheus.stages.preprocess.autoencoder import PreprocessAEStage
-from morpheus.stages.preprocess.autoencoder import TrainAEStage
+from morpheus.stages.preprocess.preprocess_ae_stage import PreprocessAEStage
+from morpheus.stages.preprocess.preprocess_ae_stage import TrainAEStage
 from morpheus.stages.preprocess.preprocessing import DeserializeStage
 from morpheus.stages.preprocess.preprocessing import DropNullStage
 from morpheus.stages.preprocess.preprocessing import PreprocessFILStage
