+++ tests/test_filter_detections_stage_pipe.py
@@ -21,7 +21,7 @@ import numpy as np
 from morpheus.pipeline import LinearPipeline
 from morpheus.stages.general_stages import FilterDetectionsStage
 from morpheus.stages.input.from_file import FileSourceStage
-from morpheus.stages.output.serialize import SerializeStage
+from morpheus.stages.postprocess.serialize import SerializeStage
 from morpheus.stages.output.to_file import WriteToFileStage
 from morpheus.stages.preprocess.preprocessing import DeserializeStage
 from utils import TEST_DIRS
