+++ tests/test_hammah.py
@@ -27,11 +27,11 @@ from morpheus.stages.general_stages import AddScoresStage
 from morpheus.stages.general_stages import MonitorStage
 from morpheus.stages.inference.inference_ae import AutoEncoderInferenceStage
 from morpheus.stages.input.from_cloudtrail import CloudTrailSourceStage
-from morpheus.stages.output.serialize import SerializeStage
+from morpheus.stages.postprocess.serialize import SerializeStage
 from morpheus.stages.output.to_file import WriteToFileStage
-from morpheus.stages.output.validation import ValidationStage
+from morpheus.stages.postprocess.validation import ValidationStage
 from morpheus.stages.postprocess.timeseries import TimeSeriesStage
-from morpheus.stages.preprocess import autoencoder
+from morpheus.stages.preprocess import preprocess_ae_stage
 from utils import TEST_DIRS
 from utils import calc_error_val
 
@@ -40,7 +40,7 @@ from utils import calc_error_val
 
 @pytest.mark.slow
 @pytest.mark.use_python
-@pytest.mark.reload_modules(autoencoder)
+@pytest.mark.reload_modules(preprocess_ae_stage)
 @pytest.mark.usefixtures("reload_modules")
 @mock.patch('morpheus.pipeline.preprocess.autoencoder.AutoEncoder')
 def test_hammah_roleg(mock_ae, config, tmp_path):
@@ -79,8 +79,8 @@ def test_hammah_roleg(mock_ae, config, tmp_path):
 
     pipe = LinearPipeline(config)
     pipe.set_source(CloudTrailSourceStage(config, input_glob=input_glob, sort_glob=True))
-    pipe.add_stage(autoencoder.TrainAEStage(config, train_data_glob=train_data_glob, seed=42, sort_glob=True))
-    pipe.add_stage(autoencoder.PreprocessAEStage(config))
+    pipe.add_stage(preprocess_ae_stage.TrainAEStage(config, train_data_glob=train_data_glob, seed=42, sort_glob=True))
+    pipe.add_stage(preprocess_ae_stage.PreprocessAEStage(config))
     pipe.add_stage(AutoEncoderInferenceStage(config))
     pipe.add_stage(AddScoresStage(config))
     pipe.add_stage(
@@ -114,7 +114,7 @@ def test_hammah_roleg(mock_ae, config, tmp_path):
 
 @pytest.mark.slow
 @pytest.mark.use_python
-@pytest.mark.reload_modules(autoencoder)
+@pytest.mark.reload_modules(preprocess_ae_stage)
 @pytest.mark.usefixtures("reload_modules")
 @mock.patch('morpheus.pipeline.preprocess.autoencoder.AutoEncoder')
 def test_hammah_user123(mock_ae, config, tmp_path):
@@ -152,8 +152,8 @@ def test_hammah_user123(mock_ae, config, tmp_path):
 
     pipe = LinearPipeline(config)
     pipe.set_source(CloudTrailSourceStage(config, input_glob=input_glob, sort_glob=True))
-    pipe.add_stage(autoencoder.TrainAEStage(config, train_data_glob=train_data_glob, seed=42, sort_glob=True))
-    pipe.add_stage(autoencoder.PreprocessAEStage(config))
+    pipe.add_stage(preprocess_ae_stage.TrainAEStage(config, train_data_glob=train_data_glob, seed=42, sort_glob=True))
+    pipe.add_stage(preprocess_ae_stage.PreprocessAEStage(config))
     pipe.add_stage(AutoEncoderInferenceStage(config))
     pipe.add_stage(AddScoresStage(config))
     pipe.add_stage(
