+++ tests/test_phishing.py
@@ -26,9 +26,9 @@ from morpheus.stages.general_stages import AddClassificationsStage
 from morpheus.stages.general_stages import MonitorStage
 from morpheus.stages.inference.inference_triton import TritonInferenceStage
 from morpheus.stages.input.from_file import FileSourceStage
-from morpheus.stages.output.serialize import SerializeStage
+from morpheus.stages.postprocess.serialize import SerializeStage
 from morpheus.stages.output.to_file import WriteToFileStage
-from morpheus.stages.output.validation import ValidationStage
+from morpheus.stages.postprocess.validation import ValidationStage
 from morpheus.stages.preprocess.preprocessing import DeserializeStage
 from morpheus.stages.preprocess.preprocessing import PreprocessNLPStage
 from utils import TEST_DIRS
