+++ tests/test_serialize_stage_pipe.py
@@ -23,7 +23,7 @@ from morpheus.pipeline import LinearPipeline
 from morpheus.pipeline.file_types import FileTypes
 from morpheus.stages.input.from_file import FileSourceStage
 from morpheus.stages.input.utils import read_file_to_df
-from morpheus.stages.output.serialize import SerializeStage
+from morpheus.stages.postprocess.serialize import SerializeStage
 from morpheus.stages.output.to_file import WriteToFileStage
 from morpheus.stages.preprocess.preprocessing import DeserializeStage
 from utils import TEST_DIRS
